rename: 컴포넌트 폴더 통합
